attendance_backend/requirements.txt
# Django and core dependencies
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
django-filter>=23.0,<26.0

# Production server
gunicorn>=21.0,<23.0

# Database (for production - PostgreSQL)
psycopg2-binary>=2.9,<3.0

# Environment variables
python-decouple>=3.8,<4.0

# For timezone handling
pytz>=2023.0

# Development tools
pytest>=7.0,<9.0
pytest-django>=4.5,<5.0
coverage>=7.0,<8.0
